This commit was manufactured by cvs2svn to create tag 'release-2-6-2'.

+++ source/tools/toolutil/uparse.c
@@ -385,7 +385,7 @@ u_parseUTF8(const char *source, int32_t sLen, char *dest, int32_t destCapacity, 
     int32_t i = 0;
     unsigned int value = 0;
     if(sLen == -1) {
-        sLen = (int32_t)strlen(source);
+        sLen = strlen(source);
     }
     
     while(read < source+sLen) {
